requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# OpenEnv
openenv-core==0.1.0

# LLM client (for inference.py)
openai==1.35.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1